Add support for inheritance
(I made your test work ;-))

This is probably a breaking change for people that created custom FluentValidationRules, as I changed the RuleContext.

Because we have to recursively go through all the inherited schemas, the "Apply" method of FluentValidationRule will now be called for each individual schema, so I had to add that schema to the RuleContext (otherwise the rules themselves would have to do that recursion, which seems not the right way to handle this)

I also changed all the existing rules so they now look at "context.Schema" instead of "context.SchemaProcessorContext.Schema".
Not sure if we still need to pass the schemaProcessorContext... there's none of the built in rules that depend on it... of course custom ones might...

I also updated the packages (there was a "package with vulnerability" warning), and also got rid of the other warnings ;-)